+++ libraries/base/changelog.md
@@ -5,6 +5,10 @@
 
   * Add `Alternative` instance for `ZipList` (#13520)
 
+  * Add instances `Num`, `Functor`, `Applicative`, `Monad`, `Semigroup`
+    and `Monoid` for `Data.Ord.Down` (#13097).
+
+
 ## 4.10.0.0 *April 2017*
   * Bundled with GHC *TBA*